@@ -1,172 +1,76 @@
 <?php
 
-Auth::routes();
+use App\Http\Controllers\DemoController;
+use App\Http\Controllers\Menu\MenuGroupController;
+use App\Http\Controllers\Menu\MenuItemController;
+use App\Http\Controllers\RoleAndPermission\AssignPermissionController;
+use App\Http\Controllers\RoleAndPermission\AssignUserToRoleController;
+use App\Http\Controllers\RoleAndPermission\ExportPermissionController;
+use App\Http\Controllers\RoleAndPermission\ExportRoleController;
+use App\Http\Controllers\RoleAndPermission\ImportPermissionController;
+use App\Http\Controllers\RoleAndPermission\ImportRoleController;
+use App\Http\Controllers\RoleAndPermission\PermissionController;
+use App\Http\Controllers\RoleAndPermission\RoleController;
+use Illuminate\Support\Facades\Route;
+use App\Models\User;
+use App\Http\Controllers\UserController;
+use App\Models\Category;
 
-//Route::get('/test', 'TestController@index')->name('test');
-Route::get('/privacy-policy', 'HomeController@privacy_policy')->name('privacy_policy');
-Route::get('/terms-of-use', 'HomeController@terms_of_use')->name('terms_of_use');
+/*
+|--------------------------------------------------------------------------
+| Web Routes
+|--------------------------------------------------------------------------
+|
+| Here is where you can register web routes for your application. These
+| routes are loaded by the RouteServiceProvider within a group which
+| contains the "web" middleware group. Now create something great!
+|
+*/
 
-
-Route::group(['middleware' => 'auth'], function () {
-
-    Route::get('/', 'HomeController@dashboard')->name('home');
-    Route::get('/home', 'HomeController@dashboard')->name('home');
-    Route::get('/dashboard', 'HomeController@dashboard')->name('dashboard');
-
-    Route::group(['prefix' => 'my_account'], function() {
-        Route::get('/', 'MyAccountController@edit_profile')->name('my_account');
-        Route::put('/', 'MyAccountController@update_profile')->name('my_account.update');
-        Route::put('/change_password', 'MyAccountController@change_pass')->name('my_account.change_pass');
-    });
-
-    /*************** Support Team *****************/
-    Route::group(['namespace' => 'SupportTeam',], function(){
-
-        /*************** Students *****************/
-        Route::group(['prefix' => 'students'], function(){
-            Route::get('reset_pass/{st_id}', 'StudentRecordController@reset_pass')->name('st.reset_pass');
-            Route::get('graduated', 'StudentRecordController@graduated')->name('students.graduated');
-            Route::put('not_graduated/{id}', 'StudentRecordController@not_graduated')->name('st.not_graduated');
-            Route::get('list/{class_id}', 'StudentRecordController@listByClass')->name('students.list')->middleware('teamSAT');
-
-            /* Promotions */
-            Route::post('promote_selector', 'PromotionController@selector')->name('students.promote_selector');
-            Route::get('promotion/manage', 'PromotionController@manage')->name('students.promotion_manage');
-            Route::delete('promotion/reset/{pid}', 'PromotionController@reset')->name('students.promotion_reset');
-            Route::delete('promotion/reset_all', 'PromotionController@reset_all')->name('students.promotion_reset_all');
-            Route::get('promotion/{fc?}/{fs?}/{tc?}/{ts?}', 'PromotionController@promotion')->name('students.promotion');
-            Route::post('promote/{fc}/{fs}/{tc}/{ts}', 'PromotionController@promote')->name('students.promote');
-
-        });
-
-        /*************** Users *****************/
-        Route::group(['prefix' => 'users'], function(){
-            Route::get('reset_pass/{id}', 'UserController@reset_pass')->name('users.reset_pass');
-        });
-
-        /*************** TimeTables *****************/
-        Route::group(['prefix' => 'timetables'], function(){
-            Route::get('/', 'TimeTableController@index')->name('tt.index');
-
-            Route::group(['middleware' => 'teamSA'], function() {
-                Route::post('/', 'TimeTableController@store')->name('tt.store');
-                Route::put('/{tt}', 'TimeTableController@update')->name('tt.update');
-                Route::delete('/{tt}', 'TimeTableController@delete')->name('tt.delete');
-            });
-
-            /*************** TimeTable Records *****************/
-            Route::group(['prefix' => 'records'], function(){
-
-                Route::group(['middleware' => 'teamSA'], function(){
-                    Route::get('manage/{ttr}', 'TimeTableController@manage')->name('ttr.manage');
-                    Route::post('/', 'TimeTableController@store_record')->name('ttr.store');
-                    Route::get('edit/{ttr}', 'TimeTableController@edit_record')->name('ttr.edit');
-                    Route::put('/{ttr}', 'TimeTableController@update_record')->name('ttr.update');
-                });
-
-                Route::get('show/{ttr}', 'TimeTableController@show_record')->name('ttr.show');
-                Route::get('print/{ttr}', 'TimeTableController@print_record')->name('ttr.print');
-                Route::delete('/{ttr}', 'TimeTableController@delete_record')->name('ttr.destroy');
-
-            });
-
-            /*************** Time Slots *****************/
-            Route::group(['prefix' => 'time_slots', 'middleware' => 'teamSA'], function(){
-                Route::post('/', 'TimeTableController@store_time_slot')->name('ts.store');
-                Route::post('/use/{ttr}', 'TimeTableController@use_time_slot')->name('ts.use');
-                Route::get('edit/{ts}', 'TimeTableController@edit_time_slot')->name('ts.edit');
-                Route::delete('/{ts}', 'TimeTableController@delete_time_slot')->name('ts.destroy');
-                Route::put('/{ts}', 'TimeTableController@update_time_slot')->name('ts.update');
-            });
-
-        });
-
-        /*************** Payments *****************/
-        Route::group(['prefix' => 'payments'], function(){
-
-            Route::get('manage/{class_id?}', 'PaymentController@manage')->name('payments.manage');
-            Route::get('invoice/{id}/{year?}', 'PaymentController@invoice')->name('payments.invoice');
-            Route::get('receipts/{id}', 'PaymentController@receipts')->name('payments.receipts');
-            Route::get('pdf_receipts/{id}', 'PaymentController@pdf_receipts')->name('payments.pdf_receipts');
-            Route::post('select_year', 'PaymentController@select_year')->name('payments.select_year');
-            Route::post('select_class', 'PaymentController@select_class')->name('payments.select_class');
-            Route::delete('reset_record/{id}', 'PaymentController@reset_record')->name('payments.reset_record');
-            Route::post('pay_now/{id}', 'PaymentController@pay_now')->name('payments.pay_now');
-        });
-
-        /*************** Pins *****************/
-        Route::group(['prefix' => 'pins'], function(){
-            Route::get('create', 'PinController@create')->name('pins.create');
-            Route::get('/', 'PinController@index')->name('pins.index');
-            Route::post('/', 'PinController@store')->name('pins.store');
-            Route::get('enter/{id}', 'PinController@enter_pin')->name('pins.enter');
-            Route::post('verify/{id}', 'PinController@verify')->name('pins.verify');
-            Route::delete('/', 'PinController@destroy')->name('pins.destroy');
-        });
-
-        /*************** Marks *****************/
-        Route::group(['prefix' => 'marks'], function(){
-
-           // FOR teamSA
-            Route::group(['middleware' => 'teamSA'], function(){
-                Route::get('batch_fix', 'MarkController@batch_fix')->name('marks.batch_fix');
-                Route::put('batch_update', 'MarkController@batch_update')->name('marks.batch_update');
-                Route::get('tabulation/{exam?}/{class?}/{sec_id?}', 'MarkController@tabulation')->name('marks.tabulation');
-                Route::post('tabulation', 'MarkController@tabulation_select')->name('marks.tabulation_select');
-                Route::get('tabulation/print/{exam}/{class}/{sec_id}', 'MarkController@print_tabulation')->name('marks.print_tabulation');
-            });
-
-            // FOR teamSAT
-            Route::group(['middleware' => 'teamSAT'], function(){
-                Route::get('/', 'MarkController@index')->name('marks.index');
-                Route::get('manage/{exam}/{class}/{section}/{subject}', 'MarkController@manage')->name('marks.manage');
-                Route::put('update/{exam}/{class}/{section}/{subject}', 'MarkController@update')->name('marks.update');
-                Route::put('comment_update/{exr_id}', 'MarkController@comment_update')->name('marks.comment_update');
-                Route::put('skills_update/{skill}/{exr_id}', 'MarkController@skills_update')->name('marks.skills_update');
-                Route::post('selector', 'MarkController@selector')->name('marks.selector');
-                Route::get('bulk/{class?}/{section?}', 'MarkController@bulk')->name('marks.bulk');
-                Route::post('bulk', 'MarkController@bulk_select')->name('marks.bulk_select');
-            });
-
-            Route::get('select_year/{id}', 'MarkController@year_selector')->name('marks.year_selector');
-            Route::post('select_year/{id}', 'MarkController@year_selected')->name('marks.year_select');
-            Route::get('show/{id}/{year}', 'MarkController@show')->name('marks.show');
-            Route::get('print/{id}/{exam_id}/{year}', 'MarkController@print_view')->name('marks.print');
-
-        });
-
-        Route::resource('students', 'StudentRecordController');
-        Route::resource('users', 'UserController');
-        Route::resource('classes', 'MyClassController');
-        Route::resource('sections', 'SectionController');
-        Route::resource('subjects', 'SubjectController');
-        Route::resource('grades', 'GradeController');
-        Route::resource('exams', 'ExamController');
-        Route::resource('dorms', 'DormController');
-        Route::resource('payments', 'PaymentController');
-
-    });
-
-    /************************ AJAX ****************************/
-    Route::group(['prefix' => 'ajax'], function() {
-        Route::get('get_lga/{state_id}', 'AjaxController@get_lga')->name('get_lga');
-        Route::get('get_class_sections/{class_id}', 'AjaxController@get_class_sections')->name('get_class_sections');
-        Route::get('get_class_subjects/{class_id}', 'AjaxController@get_class_subjects')->name('get_class_subjects');
-    });
-
+Route::get('/', function () {
+    return view('auth/login');
 });
 
-/************************ SUPER ADMIN ****************************/
-Route::group(['namespace' => 'SuperAdmin','middleware' => 'super_admin', 'prefix' => 'super_admin'], function(){
+Route::group(['middleware' => ['auth', 'verified']], function () {
+    Route::get('/dashboard', function () {
+        return view('home', ['users' => User::get(),]);
+    });
+    //user list
 
-    Route::get('/settings', 'SettingController@index')->name('settings');
-    Route::put('/settings', 'SettingController@update')->name('settings.update');
+    Route::prefix('user-management')->group(function () {
+        Route::resource('user', UserController::class);
+        Route::post('import', [UserController::class, 'import'])->name('user.import');
+        Route::get('export', [UserController::class, 'export'])->name('user.export');
+        Route::get('demo', DemoController::class)->name('user.demo');
+    });
 
-});
+    Route::prefix('menu-management')->group(function () {
+        Route::resource('menu-group', MenuGroupController::class);
+        Route::resource('menu-item', MenuItemController::class);
+    });
+    Route::group(['prefix' => 'role-and-permission'], function () {
+        //role
+        Route::resource('role', RoleController::class);
+        Route::get('role/export', ExportRoleController::class)->name('role.export');
+        Route::post('role/import', ImportRoleController::class)->name('role.import');
 
-/************************ PARENT ****************************/
-Route::group(['namespace' => 'MyParent','middleware' => 'my_parent',], function(){
+        //permission
+        Route::resource('permission', PermissionController::class);
+        Route::get('permission/export', ExportPermissionController::class)->name('permission.export');
+        Route::post('permission/import', ImportPermissionController::class)->name('permission.import');
 
-    Route::get('/my_children', 'MyController@children')->name('my_children');
+        //assign permission
+        Route::get('assign', [AssignPermissionController::class, 'index'])->name('assign.index');
+        Route::get('assign/create', [AssignPermissionController::class, 'create'])->name('assign.create');
+        Route::get('assign/{role}/edit', [AssignPermissionController::class, 'edit'])->name('assign.edit');
+        Route::put('assign/{role}', [AssignPermissionController::class, 'update'])->name('assign.update');
+        Route::post('assign', [AssignPermissionController::class, 'store'])->name('assign.store');
 
+        //assign user to role
+        Route::get('assign-user', [AssignUserToRoleController::class, 'index'])->name('assign.user.index');
+        Route::get('assign-user/create', [AssignUserToRoleController::class, 'create'])->name('assign.user.create');
+        Route::post('assign-user', [AssignUserToRoleController::class, 'store'])->name('assign.user.store');
+        Route::get('assing-user/{user}/edit', [AssignUserToRoleController::class, 'edit'])->name('assign.user.edit');
+        Route::put('assign-user/{user}', [AssignUserToRoleController::class, 'update'])->name('assign.user.update');
+    });
 });
