@@ -2,39 +2,46 @@
 
 namespace Database\Factories;
 
-use App\Helpers\Qs;
-use App\User;
+use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 
 class UserFactory extends Factory
 {
-/**
-* The name of the factory's corresponding model.
-*
-* @var string
-*/
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
     protected $model = User::class;
 
-/**
-* Define the model's default state.
-*
-* @return array
-*/
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
     public function definition()
     {
-        // Create random User Type
-        $user_type = Qs::getStaff(['super_admin', 'librarian'])[rand(0,2)];
-
         return [
-            'name' => $this->faker->name,
-            'email' => $this->faker->safeEmail,
-            'username' => $this->faker->userName,
-            'password' => Hash::make($user_type),
-            'user_type' => $user_type,
-            'code' => strtoupper(Str::random(10)),
+            'name' => $this->faker->name(),
+            'email' => $this->faker->unique()->safeEmail(),
+            'email_verified_at' => now(),
+            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
             'remember_token' => Str::random(10),
         ];
+    }
+
+    /**
+     * Indicate that the model's email address should be unverified.
+     *
+     * @return \Illuminate\Database\Eloquent\Factories\Factory
+     */
+    public function unverified()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'email_verified_at' => null,
+            ];
+        });
     }
 }
