@@ -1,11 +1,10 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
-use App\Helpers\Qs;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
@@ -15,22 +14,9 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->id();
             $table->string('name');
-            $table->string('email', 100)->unique()->nullable();
-            $table->string('code', 100)->unique();
-            $table->string('username', 100)->nullable()->unique();
-            $table->string('user_type');
-            $table->string('dob')->nullable();
-            $table->string('gender')->nullable();
-            $table->string('photo')->default(Qs::getDefaultUserImage());
-            $table->string('phone')->nullable();
-            $table->string('phone2')->nullable();
-            $table->unsignedInteger('bg_id')->nullable();
-            $table->unsignedInteger('state_id')->nullable();
-            $table->unsignedInteger('lga_id')->nullable();
-            $table->unsignedInteger('nal_id')->nullable();
-            $table->string('address')->nullable();
+            $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
@@ -47,4 +33,4 @@ class CreateUsersTable extends Migration
     {
         Schema::dropIfExists('users');
     }
-}
+};
