@@ -1,6 +1,8 @@
-document.addEventListener('DOMContentLoaded', function() {
+/**
+ *
+ * You can write your JS code here, DO NOT touch the default style file
+ * because it will make it harder for you to update.
+ *
+ */
 
-    $('.date-pick').datepicker();
-
-});
-
+"use strict";
